+++ go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/zitadel/oidc/v3 v3.45.5
 	github.com/zitadel/zitadel v1.80.0-v2.20.0.20250619094244-3a4298c1794a
 	github.com/zitadel/zitadel-go/v3 v3.28.0
-	go.miloapis.com/milo v0.24.6
+	go.miloapis.com/milo v0.24.10
 	go.uber.org/zap v1.27.1
 	golang.org/x/oauth2 v0.35.0
 	golang.org/x/sync v0.19.0
