+++ internal/controller/serviceaccount_controller_test.go
@@ -20,7 +20,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 )
 
-var _ = Describe("MachineAccount Controller", func() {
+var _ = Describe("ServiceAccount Controller", func() {
 	Context("When reconciling a resource", func() {
 
 		It("should successfully reconcile the resource", func() {
